refactor(suite): give each Sign & verify tab its own form
The page state moves up into `SignVerify`, which now only picks the shell and
keys the form on it. Each tab therefore mounts its own form and takes its state
with it when you leave, so the outcome no longer has to be cleared by hand — the
effect that watched the tab is gone.
